add cdc data migration framework
