+++ pkg/migrate/migrate.go
@@ -16,22 +16,17 @@ package migrate
 import (
 	"context"
 	"fmt"
+	"sort"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/log"
-	"github.com/pingcap/tiflow/cdc/model"
 	"github.com/pingcap/tiflow/pkg/config"
 	cerror "github.com/pingcap/tiflow/pkg/errors"
 	"github.com/pingcap/tiflow/pkg/etcd"
-	"github.com/pingcap/tiflow/pkg/pdutil"
 	"github.com/pingcap/tiflow/pkg/security"
-	"github.com/pingcap/tiflow/pkg/txnutil/gc"
 	pd "github.com/tikv/pd/client"
-	clientV3 "go.etcd.io/etcd/client/v3"
-	"go.etcd.io/etcd/client/v3/concurrency"
 	"go.uber.org/atomic"
 	"go.uber.org/zap"
 	"google.golang.org/grpc"
@@ -39,8 +34,6 @@ import (
 )
 
 const (
-	// cdcMetaVersion is hard code value indicate the metaVersion of TiCDC
-	cdcMetaVersion          = 1
 	etcdSessionTTL          = 10
 	campaignTimeoutDuration = 1 * time.Minute
 	noMetaVersion           = -1
@@ -49,12 +42,6 @@ const (
 	oldChangefeedPrefix     = "/tidb/cdc/changefeed/info"
 	oldGcServiceID          = "ticdc"
 )
-
-type keys map[string]string
-
-func (k keys) addPair(old, new string) {
-	k[old] = new
-}
 
 // Migrator migrates the cdc metadata
 type Migrator interface {
@@ -74,13 +61,8 @@ type migrator struct {
 	// oldMetaVersion int
 	newMetaVersion int
 	metaVersionKey string
-	// cdc old owner key
-	oldOwnerKey string
 	// etcd client
 	cli *etcd.CDCEtcdClient
-	// all keyPrefixes needed to be migrated or update
-	// map from oldKeyPrefix to newKeyPrefix
-	keyPrefixes keys
 
 	done atomic.Bool
 
@@ -103,12 +85,11 @@ func NewMigrator(cli *etcd.CDCEtcdClient,
 	pdEndpoints []string,
 	serverConfig *config.ServerConfig,
 ) Migrator {
+	sort.Sort(byVersion(migrationRecords))
 	return &migrator{
-		newMetaVersion:                cdcMetaVersion,
+		newMetaVersion:                migrationRecords[len(migrationRecords)-1].version,
 		metaVersionKey:                etcd.DefaultClusterAndMetaPrefix + etcd.MetaVersionKey,
-		oldOwnerKey:                   "/ticdc/cdc/owner",
 		cli:                           cli,
-		keyPrefixes:                   make(keys),
 		pdEndpoints:                   pdEndpoints,
 		config:                        serverConfig,
 		createPDClientFunc:            createPDClient,
@@ -124,6 +105,91 @@ func (m *migrator) MarkMigrateDone() {
 // IsMigrateDone check if migration is done
 func (m *migrator) IsMigrateDone() bool {
 	return m.done.Load()
+}
+
+// Migrate migrate etcd meta data
+func (m *migrator) Migrate(ctx context.Context) error {
+	version, err := getMetaVersion(ctx, m.cli.Client, m.cli.ClusterID)
+	if err != nil {
+		return errors.Trace(err)
+	}
+
+	shouldMigrate := false
+	oldVersion, newVersion := 0, m.newMetaVersion
+
+	if version == noMetaVersion {
+		shouldMigrate = true
+	} else {
+		oldVersion = version
+		shouldMigrate = oldVersion < newVersion
+	}
+
+	if !shouldMigrate {
+		return nil
+	}
+	for _, record := range migrationRecords {
+		if record.version > oldVersion {
+			if err := record.migrate(ctx, m); err != nil {
+				log.Error("migrate failed",
+					zap.Int("version", record.version),
+					zap.Error(err))
+				return errors.Trace(err)
+			}
+			// update metaVersion
+			_, err = m.cli.Client.Put(ctx, m.metaVersionKey, fmt.Sprintf("%d", m.newMetaVersion))
+			if err != nil {
+				log.Error("update meta version failed, etcd meta data migration failed", zap.Error(err))
+				return cerror.WrapError(cerror.ErrEtcdMigrateFailed, err)
+			}
+			log.Info("etcd data migration successful",
+				zap.Int("version", record.version))
+		}
+	}
+	return nil
+}
+
+// ShouldMigrate checks if we should migrate etcd meta data
+func (m *migrator) ShouldMigrate(ctx context.Context) (bool, error) {
+	version, err := getMetaVersion(ctx, m.cli.Client, m.cli.ClusterID)
+	if err != nil {
+		return false, errors.Trace(err)
+	}
+	return version != m.newMetaVersion, nil
+}
+
+// WaitMetaVersionMatched checks and waits until the metaVersion in etcd
+// matched to lock cdcMetaVersion
+func (m *migrator) WaitMetaVersionMatched(ctx context.Context) error {
+	version, err := getMetaVersion(ctx, m.cli.Client, m.cli.ClusterID)
+	if err != nil {
+		return errors.Trace(err)
+	}
+	if version == m.newMetaVersion {
+		return nil
+	}
+
+	ticker := time.NewTicker(time.Second)
+	defer ticker.Stop()
+	warnLogTicker := time.NewTicker(migrateLogsWarnDuration)
+	defer warnLogTicker.Stop()
+	start := time.Now()
+	for {
+		select {
+		case <-ctx.Done():
+			return ctx.Err()
+		case <-ticker.C:
+			version, err := getMetaVersion(ctx, m.cli.Client, m.cli.ClusterID)
+			if err != nil {
+				return errors.Trace(err)
+			}
+			if version == m.newMetaVersion {
+				return nil
+			}
+		case <-warnLogTicker.C:
+			log.Warn("meta data migrating last too long",
+				zap.Duration("duration", time.Since(start)))
+		}
+	}
 }
 
 func createPDClient(ctx context.Context,
@@ -151,250 +217,6 @@ func createPDClient(ctx context.Context,
 		))
 }
 
-// Note: we do not use etcd transaction to migrate key
-// as it has the maximum operation limit in a single transaction.
-// So we use a double check mechanism to make sure the migration is complete.
-// 1. check and put metaVersion
-// 2. campaign old owner
-// 3. update keys
-// 4. check meta data consistency
-// 5. update metaVersion
-func (m *migrator) migrate(ctx context.Context, etcdNoMetaVersion bool, oldVersion int) error {
-	pdClient, err := m.createPDClientFunc(ctx,
-		m.pdEndpoints, m.config.Security)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	defer pdClient.Close()
-
-	upstreamID := pdClient.GetClusterID(ctx)
-	// 1.1 check metaVersion, if the metaVersion in etcd does not match
-	// m.oldMetaVersion, it means that someone has migrated the metadata
-	metaVersion, err := getMetaVersion(ctx, m.cli.Client, m.cli.ClusterID)
-	if err != nil {
-		log.Error("get meta version failed, etcd meta data migration failed", zap.Error(err))
-		return cerror.WrapError(cerror.ErrEtcdMigrateFailed, err)
-	}
-
-	if metaVersion > m.newMetaVersion {
-		log.Panic("meta version in etcd is greater than the meta version in TiCDC",
-			zap.Int("etcdMetaVersion", metaVersion), zap.Int("cdcMetaVersion", m.newMetaVersion))
-	}
-
-	// if metaVersion in etcd is equal to m.newMetaVersion,
-	// it means that there is no need to migrate
-	if !etcdNoMetaVersion && metaVersion == m.newMetaVersion {
-		log.Warn("meta version no match, no need to migrate")
-		return nil
-	}
-
-	// 1.2 put metaVersionKey to etcd to panic old version cdc server
-	if etcdNoMetaVersion {
-		_, err := m.cli.Client.Put(ctx, m.metaVersionKey, fmt.Sprintf("%d", oldVersion))
-		if err != nil {
-			log.Error("put meta version failed, etcd meta data migration failed", zap.Error(err))
-			return cerror.WrapError(cerror.ErrEtcdMigrateFailed, err)
-		}
-	}
-
-	// 3.campaign old owner to make sure old keys will not be updates
-	campaignCtx, cancel := context.WithTimeout(ctx, campaignTimeoutDuration)
-	defer cancel()
-	if err := m.campaignOldOwner(campaignCtx); err != nil {
-		if errors.ErrorEqual(err, context.DeadlineExceeded) {
-			log.Error("campaign old owner timeout",
-				zap.Duration("duration", campaignTimeoutDuration))
-		}
-		log.Error("campaign old owner failed, etcd meta data migration failed",
-			zap.Error(err))
-		return cerror.WrapError(cerror.ErrEtcdMigrateFailed, err)
-	}
-
-	beforeKV := make(map[string][]byte)
-	// 4.campaign owner successfully, begin to migrate data
-	for oldPrefix, newPrefix := range m.keyPrefixes {
-		resp, err := m.cli.Client.Get(ctx, oldPrefix, clientV3.WithPrefix())
-		if err != nil {
-			log.Error("get old meta data failed, etcd meta data migration failed",
-				zap.Error(err))
-			return cerror.WrapError(cerror.ErrEtcdMigrateFailed, err)
-		}
-		for _, v := range resp.Kvs {
-			oldKey := string(v.Key)
-			newKey := newPrefix + oldKey[len(oldPrefix):]
-			beforeKV[newKey] = v.Value
-			log.Info("migrate key", zap.String("oldKey", oldKey), zap.String("newKey", newKey))
-			if strings.HasPrefix(string(v.Key), oldChangefeedPrefix) {
-				info := new(model.ChangeFeedInfo)
-				err = info.Unmarshal(v.Value)
-				if err != nil {
-					log.Error("unmarshal changefeed failed",
-						zap.Error(err))
-					return cerror.WrapError(cerror.ErrEtcdMigrateFailed, err)
-				}
-				info.UpstreamID = upstreamID
-				var str string
-				str, err = info.Marshal()
-				if err != nil {
-					log.Error("marshal changefeed failed",
-						zap.Error(err))
-					return cerror.WrapError(cerror.ErrEtcdMigrateFailed, err)
-				}
-				_, err = m.cli.Client.Put(ctx, newKey, str)
-			} else {
-				_, err = m.cli.Client.Put(ctx, newKey, string(v.Value))
-			}
-			if err != nil {
-				log.Error("put new meta data failed, etcd meta data migration failed",
-					zap.Error(err))
-				return cerror.WrapError(cerror.ErrEtcdMigrateFailed, err)
-			}
-		}
-	}
-
-	err = m.migrateGcServiceSafePointFunc(ctx, pdClient,
-		m.config.Security, m.cli.GetGCServiceID(), m.config.GcTTL)
-	if err != nil {
-		log.Error("update meta version failed, etcd meta data migration failed", zap.Error(err))
-		return cerror.WrapError(cerror.ErrEtcdMigrateFailed, err)
-	}
-
-	// 5. update metaVersion
-	_, err = m.cli.Client.Put(ctx, m.metaVersionKey, fmt.Sprintf("%d", m.newMetaVersion))
-	if err != nil {
-		log.Error("update meta version failed, etcd meta data migration failed", zap.Error(err))
-		return cerror.WrapError(cerror.ErrEtcdMigrateFailed, err)
-	}
-	log.Info("etcd data migration successful")
-	return nil
-}
-
-func migrateGcServiceSafePoint(ctx context.Context,
-	pdClient pd.Client,
-	config *security.Credential,
-	newGcServiceID string,
-	ttl int64,
-) error {
-	gcServiceSafePoins, err := pdutil.ListGcServiceSafePoint(ctx, pdClient,
-		config)
-	if err != nil {
-		log.Error("list gc service safepoint failed",
-			zap.Error(err))
-		return errors.Trace(err)
-	}
-	var cdcGcSafePoint *pdutil.ServiceSafePoint
-	for _, item := range gcServiceSafePoins.ServiceGCSafepoints {
-		if item.ServiceID == oldGcServiceID {
-			cdcGcSafePoint = item
-		}
-	}
-	if cdcGcSafePoint != nil {
-		_, err := gc.SetServiceGCSafepoint(ctx, pdClient, newGcServiceID,
-			ttl,
-			cdcGcSafePoint.SafePoint)
-		if err != nil {
-			log.Error("set gc service safepoint failed",
-				zap.Error(err))
-			return errors.Trace(err)
-		}
-		err = gc.RemoveServiceGCSafepoint(ctx, pdClient, oldGcServiceID)
-		if err != nil {
-			log.Warn("remove old gc safepoint failed", zap.Error(err))
-		}
-	}
-	return nil
-}
-
-func (m *migrator) campaignOldOwner(ctx context.Context) error {
-	sess, err := concurrency.NewSession(m.cli.Client.Unwrap(),
-		concurrency.WithTTL(etcdSessionTTL))
-	if err != nil {
-		return errors.Trace(err)
-	}
-	election := concurrency.NewElection(sess, m.oldOwnerKey)
-	defer func() {
-		_ = sess.Close()
-	}()
-
-	if err := election.Campaign(ctx, migrationCampaignKey); err != nil {
-		return errors.Trace(err)
-	}
-	return nil
-}
-
-// Migrate migrate etcd meta data
-func (m *migrator) Migrate(ctx context.Context) error {
-	version, err := getMetaVersion(ctx, m.cli.Client, m.cli.ClusterID)
-	if err != nil {
-		return errors.Trace(err)
-	}
-
-	shouldMigrate := false
-	oldVersion, newVersion := 0, cdcMetaVersion
-
-	if version == noMetaVersion {
-		shouldMigrate = true
-	} else {
-		oldVersion = version
-		shouldMigrate = oldVersion < newVersion
-	}
-
-	if !shouldMigrate {
-		return nil
-	}
-
-	m.keyPrefixes.addPair("/tidb/cdc/changefeed/info",
-		etcd.DefaultClusterAndNamespacePrefix+etcd.ChangefeedInfoKey)
-	m.keyPrefixes.addPair("/tidb/cdc/job",
-		etcd.DefaultClusterAndNamespacePrefix+etcd.ChangefeedStatusKey)
-
-	return m.migrate(ctx, version == noMetaVersion, oldVersion)
-}
-
-// ShouldMigrate checks if we should migrate etcd meta data
-func (m *migrator) ShouldMigrate(ctx context.Context) (bool, error) {
-	version, err := getMetaVersion(ctx, m.cli.Client, m.cli.ClusterID)
-	if err != nil {
-		return false, errors.Trace(err)
-	}
-	return version != cdcMetaVersion, nil
-}
-
-// WaitMetaVersionMatched checks and waits until the metaVersion in etcd
-// matched to lock cdcMetaVersion
-func (m *migrator) WaitMetaVersionMatched(ctx context.Context) error {
-	version, err := getMetaVersion(ctx, m.cli.Client, m.cli.ClusterID)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	if version == cdcMetaVersion {
-		return nil
-	}
-
-	ticker := time.NewTicker(time.Second)
-	defer ticker.Stop()
-	warnLogTicker := time.NewTicker(migrateLogsWarnDuration)
-	defer warnLogTicker.Stop()
-	start := time.Now()
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case <-ticker.C:
-			version, err := getMetaVersion(ctx, m.cli.Client, m.cli.ClusterID)
-			if err != nil {
-				return errors.Trace(err)
-			}
-			if version == cdcMetaVersion {
-				return nil
-			}
-		case <-warnLogTicker.C:
-			log.Warn("meta data migrating last too long",
-				zap.Duration("duration", time.Since(start)))
-		}
-	}
-}
-
 func getMetaVersion(ctx context.Context, cli *etcd.Client, clusterID string) (int, error) {
 	key := etcd.CDCKey{Tp: etcd.CDCKeyTypeMetaVersion, ClusterID: clusterID}
 	resp, err := cli.Get(ctx, key.String())
@@ -411,31 +233,4 @@ func getMetaVersion(ctx context.Context, cli *etcd.Client, clusterID string) (in
 		return 0, errors.Trace(err)
 	}
 	return version, nil
-}
-
-// NoOpMigrator do nothing
-type NoOpMigrator struct{}
-
-// ShouldMigrate checks if we need to migrate metadata
-func (f *NoOpMigrator) ShouldMigrate(ctx context.Context) (bool, error) {
-	return false, nil
-}
-
-// Migrate migrates the cdc metadata
-func (f *NoOpMigrator) Migrate(ctx context.Context) error {
-	return nil
-}
-
-// WaitMetaVersionMatched wait util migration is done
-func (f *NoOpMigrator) WaitMetaVersionMatched(ctx context.Context) error {
-	return nil
-}
-
-// MarkMigrateDone marks migration is done
-func (f *NoOpMigrator) MarkMigrateDone() {
-}
-
-// IsMigrateDone check if migration is done
-func (f *NoOpMigrator) IsMigrateDone() bool {
-	return true
 }
